requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.9.0
pydantic-settings==2.5.2
email-validator==2.1.0.post1
pytest
httpx
